fix: split context into separate files
