+++ client/src/app/components/AIModelDetails/index.ts
@@ -1,0 +1,2 @@
+export * from "./AIModelDetails";
+export * from "./utils";
