+++ client/src/app/components/NotificationsContext/NotificationsProvider.tsx
@@ -1,34 +1,14 @@
 import * as React from "react";
 
-import type { AlertProps } from "@patternfly/react-core";
-
-export type INotification = {
-  title: string;
-  variant: AlertProps["variant"];
-  message?: React.ReactNode;
-  hideCloseButton?: boolean;
-  timeout?: number | boolean;
-};
+import { INotification, NotificationsContext } from "./NotificationsContext";
 
 interface INotificationsProvider {
   children: React.ReactNode;
 }
 
-interface INotificationsContext {
-  pushNotification: (notification: INotification) => void;
-  dismissNotification: (key: string) => void;
-  notifications: INotification[];
-}
-
-const appContextDefaultValue = {} as INotificationsContext;
-
 const notificationDefault: Pick<INotification, "hideCloseButton"> = {
   hideCloseButton: false,
 };
-
-export const NotificationsContext = React.createContext<INotificationsContext>(
-  appContextDefaultValue,
-);
 
 export const NotificationsProvider: React.FunctionComponent<
   INotificationsProvider
