+++ client/src/app/components/NotificationsContext/index.ts
@@ -1,0 +1,3 @@
+export * from "./NotificationsContext";
+export * from "./NotificationsProvider";
+
