+++ client/src/app/components/ReadOnlyContext/ReadOnlyContext.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface IReadOnlyContext {
+  isReadOnly: boolean;
+  isLoading: boolean;
+}
+
+export const ReadOnlyContext = React.createContext<IReadOnlyContext>({
+  isReadOnly: false,
+  isLoading: true,
+});
