+++ client/src/app/pages/sbom-upload/sbom-upload.tsx
@@ -1,4 +1,4 @@
-import type React from "react";
+import React from "react";
 import { Link } from "react-router-dom";
 
 import type { AxiosError, AxiosResponse } from "axios";
@@ -15,13 +15,13 @@ import {
 import LockIcon from "@patternfly/react-icons/dist/esm/icons/lock-icon";
 
 import { DocumentMetadata } from "@app/components/DocumentMetadata";
-import { useReadOnlyContext } from "@app/components/ReadOnlyContext";
+import { ReadOnlyContext } from "@app/components/ReadOnlyContext";
 import { UploadFiles } from "@app/components/UploadFiles";
 import { useUploadSBOM } from "@app/queries/sboms";
 import { Paths } from "@app/Routes";
 
 export const SbomUpload: React.FC = () => {
-  const { isReadOnly } = useReadOnlyContext();
+  const { isReadOnly } = React.useContext(ReadOnlyContext);
   const { uploads, handleUpload, handleRemoveUpload } = useUploadSBOM();
 
   return (
